backend/requirements.txt
Flask==2.3.3
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
SQLAlchemy==2.0.23
marshmallow==3.20.1
python-dotenv==1.0.0
pytest==7.4.3
gunicorn==21.2.0
pymysql==1.1.0 